requirements.txt
requests>=2.25.0
bs4>=0.0.1
python-dotenv>=1.0.0
typing-extensions>=4.0.0
importlib-metadata>=4.6.3
beautifulsoup4>=4.9.3

# Development dependencies
pytest>=6.1.2
flake8>=3.8.4
pre-commit>=2.8.2
pytest-cov>=2.10.1
requests-mock>=1.8.0
lxml>=4.6.2
mypy>=0.790
pytest-mock>=3.3.1
